fix reverse duplicates and fix metric and local join implemented

+++ Metrics/metrics.c
@@ -6,7 +6,7 @@
 double l2(Pointer *x, Pointer *y, int d) {
     double result = 0.0, value;
     for(int i = 0; i < d; i++) {
-        value = fabs(*(double*)x[i]- *(double*)y[i]);
+        value = fabs(*(float*)x[i]- *(float*)y[i]);
         result += pow(value, 2.0);
     }
     return sqrt(result);
